add shorcutFieldOrder to right class
